python-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1
numpy==1.26.2
python-multipart==0.0.6
aiofiles==23.2.1
redis==5.0.1
python-dateutil==2.8.2
scipy==1.11.4
chardet==5.2.0
